Refactors(): command registration process
Modernizes the command registration process by switching to a dynamic registration based on Discord events and API interactions.

This change replaces the static command registration handler with a new event-driven system that registers commands on bot ready event. It introduces utilities to compare local and remote commands, allowing for updates and deletions. This ensures that commands are always up-to-date with the latest definitions and removes the need for manual registration.

Additionally, renames `01setActivity.ts` to `setActivity.ts` for clarity.